requirements.txt
python-telegram-bot==20.8
aiohttp==3.9.3
requests==2.31.0
